clip-generation-service/requirements-training.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
opencv-python>=4.5.3
albumentations>=1.0.3
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3
wandb>=0.12.0
Pillow>=8.3.2
scipy>=1.7.0
transformers>=4.15.0
torchaudio>=0.9.0
librosa>=0.8.1
psutil>=5.8.0
aiofiles>=0.8.0
pandas>=1.3.0
cryptography>=3.4.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5 